% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_junction_tree.R
\name{BIC_junction_tree}
\alias{BIC_junction_tree}
\title{Calculates the BIC value}
\usage{
BIC_junction_tree(cliques, separators, data, base_log = 2, ...)
}
\arguments{
\item{cliques}{A list containing the cliques of the junction tree.}

\item{separators}{A list containing the separators of the junction
tree.}

\item{data}{A data frame with observations of the variables.}

\item{base_log}{The base to use for the logarithm.}

\item{...}{Additional arguments passed to \code{loglikelihood}.}
}
\value{
The BIC value.
}
\description{
Calculates the BIC value for a graphical model where the
graph has the given junction tree.
}
\details{
The BIC is calculated as \deqn{loglik - (n_param / 2)
* log(n_obs)} where \eqn{loglik} is the value of the log-likelihood
function, \eqn{n_param} is the number of free parameters and
\eqn{n_obs} is the number of observations in data. A high value of the
BIC criterion is to be preferred.
}
\examples{

set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

cliques <- list(c("var1", "var2", "var3"),
                c("var2", "var3", "var5"),
                c("var5", "var6", "var7"),
                c("var1", "var4"),
                c("var2", "var5", "var6"))

separators <- list(c("var1"),
                   c("var2", "var3"),
                   c("var2", "var5"),
                   c("var5", "var6"))

# smooth is used to deal with zero probabilities.
BIC_junction_tree(cliques, separators, data, smooth = 0.1)
}
\seealso{
\code{\link{loglikelihood}} for calculating the
log-likelihood value and \code{\link{n_params_junction_tree}} for
determining the number of free parameters in the model.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
