% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_independence_test.R
\name{cond_independence_test}
\alias{cond_independence_test}
\title{Conditional independence test}
\usage{
cond_independence_test(var1, var2, cond = c(), data, smooth = 0)
}
\arguments{
\item{var1, var2}{Variables to test independence for.}

\item{cond}{Variables to condition on. If empty the test is just an
independence test.}

\item{data}{Data with realisations of the variables.}

\item{smooth}{Additional cell counts when estimating probabilities.
May be used to avoid zero probabilities.}
}
\value{
A list containing the following components:
\itemize{
\item\code{chi_sq_statistic} The calculated test statistic.
\item\code{df} Degrees of freedom.
\item\code{p_value} The p-value of the test.
}
}
\description{
Makes a conditional independece test based on a
likelihood ratio test.
}
\details{
The calculated test statistic is \deqn{
2 \sum n_comb * log((P(var1, var2|cond)) /
(P(var1|cond)P(var2|cond)))
}
where the sum is over all states of the variables, n_comb is the
number of times the current combination of states has been seen in
the data and the probability distributions are estimated. This
estimation is done with maximum likelihood if \code{smooth} is set
to 0. If this gives zero probabilities the \code{smooth} argument
must be used to prevent it in order to calculate the test size.

The test statistic follows assymptotically a chi-square distribution
with \eqn{(n1 - 1)(n2 - 1)n_cond} degrees of freedom, where \eqn{n1}
and \eqn{n2}
are the number of states for \code{var1} and \code{var2} and
\eqn{n_cond} is the number of configurations of the states of the
variables to be conditioned on.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5))


cond_independence_test("var1", "var4", data = data, smooth = 0.1)
cond_independence_test("var2", "var3", cond = c("var1", "var5"),
                       data = data, smooth = 0.1)

}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
