% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPT.R
\name{CPT}
\alias{CPT}
\title{Estimate conditional probability tables}
\usage{
CPT(adj_matrix, data, bayes_smooth = 0)
}
\arguments{
\item{adj_matrix}{The adjacency matrix for the DAG.}

\item{data}{The data the probabilities should be estimated from.}

\item{bayes_smooth}{The additional cell counts for
bayesian estimation of probability tables.}
}
\value{
A list of the conditional probability tables for the
bayesian network. If the \code{bayes_smooth} argument is zero,
it is the maximum likelihood estimates. Otherwise, it is bayesian
estimates.
}
\description{
Estimates the conditional probability tables for
bayesian network models, where the structure is given by an
adjacency matrix.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 50, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 50, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 50, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 50, replace = TRUE))
data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4))

adj_matrix_DAG <- matrix(c(0, 0, 0, 0,
                           1, 0, 0, 0,
                           1, 0, 0, 0,
                           0, 1, 0, 0),
                          nrow = 4)
CPT(adj_matrix_DAG, data)
CPT(adj_matrix_DAG, data, bayes_smooth = 1)
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
