% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_edges.R
\name{diff_edges}
\alias{diff_edges}
\title{Determine the number of different edges for graphs.}
\usage{
diff_edges(adj_mat_1, adj_mat_2)
}
\arguments{
\item{adj_mat_1, adj_mat_2}{Adjacency matrices for the two graphs.}
}
\value{
The number of edges where the two graphs differ.
}
\description{
Determines the number of different edges for two undirected
graphs, with no loops, represented by adjacency matrices.
}
\examples{
m1 <- matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), nrow = 3, ncol = 3)
m2_same <- m1
m2_diff <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), nrow = 3, ncol = 3)

diff_edges(m1, m2_same)
diff_edges(m1, m2_diff)

}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
