% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_edges.R
\name{diff_edges_tch}
\alias{diff_edges_tch}
\title{Determine the number of differing edges for t-cherry trees.}
\usage{
diff_edges_tch(adj_mat_1, adj_mat_2)
}
\arguments{
\item{adj_mat_1, adj_mat_2}{Adjacency matrices for the two graphs.}
}
\value{
The number of edges where the two graphs differ.
}
\description{
Determines the number of differing edges for two t-cherry
trees over the same universe represented by adjacency matrices,
i.e. the total number of edges minus the number of edges common to both
t-cherry trees.
}
\examples{
m1 <- matrix(c(0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0),
             nrow = 4, ncol = 4)
m2 <- matrix(c(0, 1, 1, 1, 1, 0, 0, 1, 1, 0, 0, 1, 1, 1, 1, 0),
             nrow = 4, ncol = 4)

diff_edges_tch(m1, m1)
diff_edges_tch(m1, m2)

}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
