% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CL.R
\name{is_acyclic}
\alias{is_acyclic}
\title{Test whether an undirected graph is acyclic}
\usage{
is_acyclic(adj_matrix)
}
\arguments{
\item{adj_matrix}{The adjacency matrix representing the graph.}
}
\value{
A logical value indicating whether the graph is acyclic.
}
\description{
Test whether an undirected graph, represented by
an adjacency matrix, is acyclic.
}
\details{
Notice that the function cannot cope with loops.
If the graph has loops, an error is returned.
}
\examples{
adj_matrix_cyclic <- matrix(c(0, 1, 1, 1,
                              1, 0, 0, 1,
                              1, 0, 0, 0,
                              1, 1, 0, 0),
                             nrow = 4)

is_acyclic(adj_matrix_cyclic)

adj_matrix_acyclic <- matrix(c(0, 0, 1, 1,
                               0, 0, 0, 1,
                               1, 0, 0, 0,
                               1, 1, 0, 0),
                              nrow = 4)

is_acyclic(adj_matrix_acyclic)
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
