% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kullback.R
\name{kullback}
\alias{kullback}
\title{Calculates Kullback-Leibler divergence.}
\usage{
kullback(target_p, approx_p)
}
\arguments{
\item{target_p}{Named array with the target distribution.}

\item{approx_p}{Named arrays with the approximate distribution.}
}
\value{
The Kullback-Leibler divergence.
}
\description{
Calculates the Kullback-Leibler divergence for two
discrete probability distributions over the same universe.
}
\details{
The Kullback-Leibler divergence between two probability
distributions p and q is calculated by the
formula \deqn{D(p||q) = \sum p(x)log(p(x)/q(x))} where
the sum is over all possible states of x.
}
\examples{
p_target <- array(c(0.1, 0.2, 0.05, 0.4, 0.005, 0.015, 0.13, 0.1),
                  dim = c(2, 2, 2),
                  dimnames = list("V1" = c("a", "b"),
                                  "V2" = c("a", "b"),
                                  "V3" = c("a", "b")))
p_approx <- array(c(0.01, 0.23, 0.06, 0.4, 0.005, 0.018,
                    0.137, 0.14),
                  dim = c(2, 2, 2),
                  dimnames = list("V1" = c("a", "b"),
                                  "V2" = c("a", "b"),
                                  "V3" = c("a", "b")))
kullback(p_target, p_approx)
}
\references{
\insertRef{TCJT}{tcherry}
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
