% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglikelihood.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Calculates the maximum log-likelihood}
\usage{
loglikelihood(cliques, separators, data, ...)
}
\arguments{
\item{cliques}{A list containing the cliques of the junction tree.}

\item{separators}{A list containing the separators of the junction
tree.}

\item{data}{A data frame with observations of the variables.}

\item{...}{Additional arguments passed to \code{weight_junction_tree}.}
}
\value{
The calculated log-likelihood value.
}
\description{
Calculates the value of the log-likelihood function for a
graphical model with the given junction tree after inserting estimated
probabilities.
}
\details{
Calculates the value as \deqn{n_obs * (weight - \sum H(A))}
 where \eqn{n_obs} is the number of observations in data, \eqn{weight}
 is the estimated weight of the junction tree and \eqn{H} is the
 estimated entropy of the variable \eqn{A}. The sum is over all
 variables in data. All estimated probabilities are maximum likelihood
 estimates unless the smooth argument has been used to deal with zero
 probabilities.
}
\examples{

set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

cliques <- list(c("var1", "var2", "var3"),
                c("var2", "var3", "var5"),
                c("var5", "var6", "var7"),
                c("var1", "var4"),
                c("var2", "var5", "var6"))

separators <- list(c("var1"),
                   c("var2", "var3"),
                   c("var2", "var5"),
                   c("var5", "var6"))

# smooth is used to deal with zero probabilities.
loglikelihood(cliques, separators, data, smooth = 0.1)
}
\seealso{
\code{\link{weight_junction_tree}} for calculating the weight
of a junction tree.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
