% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.R
\name{MI2}
\alias{MI2}
\alias{MI3}
\title{Calculates mutual information}
\usage{
MI2(x, y, smooth = 0, log_base = 2)

MI3(x, y, z, smooth = 0, log_base = 2)
}
\arguments{
\item{x, y, z}{Vectors of class character or factor.}

\item{smooth}{Additional cell counts for bayesian estimation of
probabilities.}

\item{log_base}{The base of the logarithmic function to be used.}
}
\value{
The mutual information given by a single numeric value.
}
\description{
Calculate mutual information for two or three
categorical variables.
}
\details{
The mutual information for two variables is calculated by the
formula \deqn{MI(x, y) = \sum P(x, y) log(P(x, y) / (P(x)P(y)))}
where the sum is over alle possible values of x and y.

The mutual information for three variables is calculated by the
formula \deqn{MI(x, y, z) = \sum P(x, y, z) log(P(x, y, z) / (P
(x)P(y)P(z)))} where the sum is over all possible values of x, y and
z.
}
\examples{
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- c(sample(c(1, 2), 100, replace = TRUE))
var1 <- as.character(var1)
var2 <- as.character(var2)
var3 <- as.character(var3)

MI2(var1, var2, smooth = 1)
MI2(var1, var2, smooth = 0.1, log_base = exp(1))

MI3(var1, var2, var3, smooth = 1)
MI3(var1, var2, var3, smooth = 0.1, log_base = exp(1))
}
\references{
\insertRef{TCJT}{tcherry}
\insertRef{EKTS}{tcherry}
}
\seealso{
\code{\link{MIk}} for mutual information for k variables.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
