% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_params_junction_tree.R
\name{n_params_junction_tree}
\alias{n_params_junction_tree}
\title{Determines the number of free parameters}
\usage{
n_params_junction_tree(cliques, separators, data)
}
\arguments{
\item{cliques}{A list containing the cliques of the junction tree.}

\item{separators}{A list containing the separators of the junction
tree.}

\item{data}{A data frame with observations of the variables.}
}
\value{
The number of free parameters.
}
\description{
Determines the number of free parameters in a graphical
model based on a junction tree for the graph.
}
\examples{

set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

cliques <- list(c("var1", "var2", "var3"),
                c("var2", "var3", "var5"),
                c("var5", "var6", "var7"),
                c("var1", "var4"),
                c("var2", "var5", "var6"))

separators <- list(c("var1"),
                   c("var2", "var3"),
                   c("var2", "var5"),
                   c("var5", "var6"))

n_params_junction_tree(cliques, separators, data)

}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
