% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry_CL.R
\name{tcherry_CL}
\alias{tcherry_CL}
\title{Determine t-cherry tree from a Chow-Liu tree}
\usage{
tcherry_CL(data, ...)
}
\arguments{
\item{data}{The data the tree structure should be based on.}

\item{...}{Additional arguments passed to \code{MI2} and
\code{MI3}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{adj_matrix} The adjacency matrix for the t-cherry
tree.
\item \code{cliques} A list containing the cliques (cherries) of
 the t-cherry tree.
\item \code{separators} A list containing the separators of a
junction tree for the t-cherry tree.
\item \code{MI} A data frame containing the mutual information
of all triplets.
}
}
\description{
Determine the structure of a t-cherry tree
from a Chow-Liu tree for data.
}
\details{
The algorithm for constructing the t-cherry tree from
a Chow-Liu tree is as described in \insertRef{EKTS}{tcherry}.

The algorithm is greedy in the sence that it always attempts to
use the three variables with highest mutual information as the
next cherry.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

# smooth used in both MI2 and MI3
(tch <- tcherry_CL(data, smooth = 0.1))

# For plotting
library(gRbase)
library(Rgraphviz)
tcherry_tree <- as(tch$adj_matrix, "graphNEL")
plot(tcherry_tree)

# For probability propagation
library(gRain)
model <- grain(tcherry_tree, data = data, smooth = 0.1)
querygrain(model)
}
\references{
\insertRef{EKTS}{tcherry}
}
\seealso{
\code{\link{ChowLiu}} for construction of Chow-Liu
trees and \code{\link{MI3}} for mutual information of three
variables.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
