% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcherry_complete_search.R
\name{tcherry_complete_search}
\alias{tcherry_complete_search}
\title{Determine the k'th order t-cherry tree from data with the highest
weight}
\usage{
tcherry_complete_search(data, k, ...)
}
\arguments{
\item{data}{The data the tree structure should be based on.}

\item{k}{The order of the t-cherry tree.}

\item{...}{Additional arguments passed to \code{weight_junction_tree}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{model} A list with the following components:
\itemize{
\item \code{weight} The weight of the final k'th order t-cherry tree.
\item \code{cliques} A list containing the cliques of
 the k'th order t-cherry tree.
\item \code{separators} A list containing the separators of a
junction tree for the k'th order t-cherry tree.
\item \code{adj_matrix} The adjacency matrix for the k'th order
t-cherry tree.
\item \code{n_edges} The number of edges in the resulting graph.
}
\item \code{n_models} The number of considered models.
}
}
\description{
Determine the structure of the k'th order t-cherry tree
from data with the highest weight based on a complete search.
}
\details{
The algorithm for constructing the t-cherry tree from
data is based on an atempt to minimize the Kullback-Leibler
divergence, by mazimising the weight. All possible structures are
determined and the one with the highest weight is chosen.

Note that this procedure is highly inefficient, and only suited for
small problems.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

(tch <- tcherry_complete_search(data, 5, smooth = 0.1))
}
\seealso{
\code{\link{weight_junction_tree}} for calculation of the
weight and \code{\link{k_tcherry_step}} for a more efficient but
greedy algorithm.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
