% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_edges.R
\name{thin_edges}
\alias{thin_edges}
\title{Thinning of edges in a graphical model with a triangulated graph}
\usage{
thin_edges(cliques, separators, data, alpha = 0.05, ...)
}
\arguments{
\item{cliques}{A list containing the cliques of the triangulated
graph, which should be thinned.}

\item{separators}{A list containing the separators of the junction
tree of the triangulated graph, which should be thinned.}

\item{data}{Data with realisations of the variables in the graph.}

\item{alpha}{Significance level used in the conditional
independece test.}

\item{...}{Additional arguments passed to
\code{cond_independence_test}}
}
\value{
A list containing the following components:
\itemize{
\item \code{adj_matrix} The adjacency matrix of the thinned graph.
\item \code{cliques} The cliques of the thinned graph.
\item \code{separators} The separators of the junction tree for
the thinned graph.
\item \code{n_edges} The number of edges in the resulting graph.
\item \code{n_edges_removed} The number of removed edges.
}
}
\description{
Thinning of edges in a graphical model with a
triangulated graph based on a likelihood ratio test for conditional
independence.
}
\details{
The edges in the graph are removed one by one if the respective
conditional independence test cannot be rejected. An edge is only
considered if its removal results in a new triangulated graph,
i.e. the edge is in one clique only.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

cliques <- list(c("var1", "var2", "var3"),
                c("var2", "var3", "var5"),
                c("var5", "var6", "var7"),
                c("var1", "var4"),
                c("var2", "var5", "var6"))

separators <- list(c("var1"),
                   c("var2", "var3"),
                   c("var2", "var5"),
                   c("var5", "var6"))

thin_edges(cliques, separators, data = data, alpha = 0.1,
               smooth = 0.1)
}
\seealso{
\code{\link{cond_independence_test}} for the test used.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
