% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_junction_tree.R
\name{weight_junction_tree}
\alias{weight_junction_tree}
\title{Calculates the weight of a junction tree}
\usage{
weight_junction_tree(cliques, separators, data, ...)
}
\arguments{
\item{cliques, separators}{Lists containing cliques and separators
given by variable names.}

\item{data}{A data frame or matrix containing the variables.}

\item{...}{Additional arguments passed to \code{MIk}.}
}
\value{
The weight of the junction tree.
}
\description{
Calculates the weight of a junction tree with the given
cliques and separators.
}
\details{
The weight of the junction tree is calculated
as \deqn{\sum MI(clique)-\sum MI(separator)} where the sum is over
all cliques and separators respectively.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4))
cliques <- list(c("var1", "var2", "var4"), c("var1", "var3", "var4"))
separators <- list(c("var1", "var4"))

weight_junction_tree(cliques, separators, data, smooth = 0.001)
}
\references{
\insertRef{EKTShyp}{tcherry}
}
\seealso{
\code{\link{MIk}} for mutual infomation for k variables.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
