% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_order1.R, R/increase_order2.R
\name{increase_order1}
\alias{increase_order1}
\alias{increase_order2}
\title{Determine a (k + 1)'th order t-cherry tree from a k'th order t-cherry
tree}
\usage{
increase_order1(tch_cliq, data, ...)

increase_order2(tch_cliq, data, ...)
}
\arguments{
\item{tch_cliq}{A list containing the cliques of the k'th order
t-cherry tree.}

\item{data}{The data the tree structure should be based on.}

\item{...}{Additional arguments passed to \code{MIk}.}
}
\value{
A list containing the following components:
\itemize{
\item \code{adj_matrix} The adjacency matrix for the (k + 1)'th order
t-cherry tree.
\item \code{weight} Weight of the junction tree (only for \code{increase_order2})
\item \code{cliques} A list containing the cliques (cherries) of
 the (k + 1)'th order t-cherry tree.
\item \code{separators} A list containing the separators of a
junction tree for the (k + 1)'th order t-cherry tree.
\item \code{n_edges} The number of edges in the resulting graph.
}
}
\description{
Determine the structure of a (k + 1)'th order
t-cherry tree from a k'th order t-cherry tree.
}
\details{
The algorithms for constructing the (k+1)'th order t-cherry
tree from a k'th order t-cherry tree are greedy algorithms.
\code{increase_order1} attempts to maximize the sum of
mutual information of the cliques and
\code{increase_order2} attempts to maximize the weight of
the junction tree. \code{increase_order2} is also a faster
implementation and a faster alternative to creating a (k + 1)'th order
t-cherry tree directly from data with \code{k_tcherry_step}. It is
therefore recommended to use this one, and \code{increase_order1}
is primarily kept for historical reasons.

In \code{increase_order1} the procedure is:
\itemize{
\item Starting from the k'th order t-cherry tree make a complete set
of the (k + 1) variables with highest mutual information which
satisfies
that this only adds one edge to the original graph. Remove these
(k + 1)
variables from later consideration.
\item Make a complete set of the (k + 1) variables with highest mutual
information which satisfies that this only adds one edge to the graph
and that k of the variables are in earlier created cliques of size
(k + 1). Remove these (k + 1) variables from later consideration.
\item Continue until all variables are in a clique of size (k + 1).
}

For \code{increase_order2} the procedure is to start from
the k'th order t-cherry tree and then choose the first cherry as the
(k + 1) variables with highest mutual information which satisfies that
it only adds one edge to the existing graph. This is the preliminary
t-cherry tree. Then all possible new cherries of size (k + 1) are
added
stepwise to this tree and the weight \deqn{\sum MI(clique) -
\sum MI(separator)} is calculated. The first sum is over the cliques
and the second over the separators of the junction tree of the
preliminary t-cherry tree. Again new cherries are only possible if
only one edge is added to the existing graph.
The one with the highest weight is chosen as the new preliminary
t-cherry tree, and the procedure is repeated untill all variables
has been added.
}
\examples{
set.seed(43)
var1 <- c(sample(c(1, 2), 100, replace = TRUE))
var2 <- var1 + c(sample(c(1, 2), 100, replace = TRUE))
var3 <- var1 + c(sample(c(0, 1), 100, replace = TRUE,
                        prob = c(0.9, 0.1)))
var4 <- c(sample(c(1, 2), 100, replace = TRUE))
var5 <- var2 + var3
var6 <- var1 - var4 + c(sample(c(1, 2), 100, replace = TRUE))
var7 <- c(sample(c(1, 2), 100, replace = TRUE))

data <- data.frame("var1" = as.character(var1),
                   "var2" = as.character(var2),
                   "var3" = as.character(var3),
                   "var4" = as.character(var4),
                   "var5" = as.character(var5),
                   "var6" = as.character(var6),
                   "var7" = as.character(var7))

ChowLiu_cliques <- list(c("var1", "var5"),
                        c("var2", "var5"),
                        c("var3", "var5"),
                        c("var3", "var7"),
                        c("var4", "var6"),
                        c("var5", "var6"))
# smooth used in MIk
(tch <- increase_order1(ChowLiu_cliques, data, smooth = 0.1))
(tch2 <- increase_order2(ChowLiu_cliques, data, smooth = 0.1))

# For plotting
library(gRbase)
library(Rgraphviz)
tcherry_tree <- as(tch$adj_matrix, "graphNEL")
plot(tcherry_tree)

# For probability propagation
library(gRain)
model <- grain(tcherry_tree, data = data, smooth = 0.1)
querygrain(model)
}
\seealso{
\code{\link{MIk}} for mutual information of k variables.
}
\author{
Katrine Kirkeby, \email{enir_tak@hotmail.com}

Maria Knudsen, \email{mariaknudsen@hotmail.dk}

Ninna Vihrs, \email{ninnavihrs@hotmail.dk}
}
