% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MI.R
\name{MI2}
\alias{MI2}
\title{Calculate mutual information}
\usage{
MI2(x, y, smooth = 0, log_base = 2)
}
\arguments{
\item{x, y, z}{Vectors of class character or factor.}

\item{smooth}{A single numeric non-negative value. See 'details' below.}

\item{log_base}{The base of the logarithmic function to be used.}
}
\value{
The mutual information given by a single numeric value.
}
\description{
Calculate mutual information for two or three variables.
}
\details{
The \code{smooth} argument is added to all cell counts in tables before normalisation. This is meant as a way to avoid zero counts, leading to zero probabilities.

    The mutual information for two variables is calculated by the formula \deqn{MI(x,y)=\sum P(x,y)log(P(x,y)/P(x)P(y))} where the sum is over alle possible values of x and y.
}
